// app/types/notion.type.ts
export enum BlockTypes {
  "bookmark" = "bookmark",
  "breadcrumb" = "breadcrumb",
  "bulleted_list_item" = "bulleted_list_item",
  "callout" = "callout",
  "child_database" = "child_database",
  "child_page" = "child_page",
  "code" = "code",
  "column" = "column",
  "column_list" = "column_list",
  "divider" = "divider",
  "embed" = "embed",
  "equation" = "equation",
  "file" = "file",
  "heading_1" = "heading_1",
  "heading_2" = "heading_2",
  "heading_3" = "heading_3",
  "image" = "image",
  "link_preview" = "link_preview",
  "link_to_page" = "link_to_page",
  "numbered_list_item" = "numbered_list_item",
  "paragraph" = "paragraph",
  "pdf" = "pdf",
  "quote" = "quote",
  "synced_block" = "synced_block",
  "table" = "table",
  "table_of_contents" = "table_of_contents",
  "table_row" = "table_row",
  "template" = "template",
  "to_do" = "to_do",
  "toggle" = "toggle",
  "unsupported" = "unsupported",
  "video" = "video",
}

/**
 * Notion Block
 * https://developers.notion.com/reference/block
 */
export interface NotionBasicBlockDoc {
  object: "block";
  id: string;
  parent: {
    type: "page_id";
    page_id: string;
  };
  created_time: string;
  last_edited_time: string;
  created_by: {
    object: string;
    id: string;
  };
  last_edited_by: {
    object: string;
    id: string;
  };
  has_children: boolean;
  archived: boolean;
  [key: string]: any;
  type: BlockTypes;
}
// ...
enum NotionRichTextType {
  "text",
  "mention", // https://developers.notion.com/reference/rich-text#mention
  "equation", // https://developers.notion.com/reference/rich-text#equation
}

export interface NotionRichText {
  type: NotionRichTextType;
  annotations: NotionRichTextAnnotation;
  plain_text: string; // "Some words "
  href?: string;
}

export interface NotionTextTypeRichText extends NotionRichText {
  text: NotionRichTextText;
}

/**
 * https://developers.notion.com/reference/rich-text#text
 */
export interface NotionRichTextText {
  content: string;
  link?: { url: string } | null;
}

/**
 * https://developers.notion.com/reference/rich-text#the-annotation-object
 */
export interface NotionRichTextAnnotation {
  bold: boolean; // false
  italic: boolean; // false
  strikethrough: boolean; // false
  underline: boolean; // false
  code: boolean; // false
  color: string; // default"
}